@@ -9,7 +9,7 @@
         }
     },
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.0",
         "symfony/framework-bundle": "^3.4 | ^4.0",
         "symfony/form": "^3.4 | ^4.0",
         "symfony/templating": "^3.4 | ^4.0",
@@ -17,19 +17,22 @@
         "symfony/validator": "^3.4 | ^4.0",
         "symfony/serializer": "^3.4 | ^4.0",
         "symfony/cache": "^3.4 | ^4.0",
-        "commercetools/php-sdk": "^2.5",
+        "commercetools/php-sdk": "dev-develop",
         "symfony/var-dumper": "^3.4 | ^4.0",
         "symfony/dotenv": "^3.4 | ^4.0",
         "symfony/stopwatch": "^3.4 | ^4.0",
         "symfony/workflow": "^3.4 | ^4.0",
-        "twig/extensions": "^1.5"
+        "twig/extensions": "^1.5",
+        "symfony/console": "^3.4 | ^4.0",
+        "symfony/twig-bundle": "^3.4 | ^4.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^7.0 | ^6.0 | ^5.0",
         "symfony/phpunit-bridge": "^3.4 | ^4.0",
         "commercetools/commercetools-api-reference": "dev-master",
         "matthiasnoback/symfony-dependency-injection-test": "^3.0",
-        "symfony/test-pack": "^1.0"
+        "symfony/test-pack": "^1.0",
+        "cache/filesystem-adapter": "^1.0"
     },
     "extra": {
         "branch-alias": {
