@@ -12,13 +12,25 @@ use Symfony\Component\HttpKernel\DataCollector\DataCollector;
 
 class CommercetoolsDataCollector extends DataCollector
 {
+    /**
+     * @var Profile
+     */
     private $profile;
 
+    /**
+     * CommercetoolsDataCollector constructor.
+     * @param Profile $profile
+     */
     public function __construct(Profile $profile)
     {
         $this->profile = $profile;
     }
 
+    /**
+     * @param Request $request
+     * @param Response $response
+     * @param \Exception|null $exception
+     */
     public function collect(Request $request, Response $response, \Exception $exception = null)
     {
         $this->data['duration'] = $this->profile->getDuration();
@@ -27,17 +39,17 @@ class CommercetoolsDataCollector extends DataCollector
 
     public function getDuration()
     {
-        return isset($this->data['duration']) ? $this->data['duration'] : 0;
+        return $this->data['duration'] ?? 0;
     }
 
     public function getRequestCount()
     {
-        return count($this->data['requestInfos']);
+        return isset($this->data['requestInfos']) ? count($this->data['requestInfos']) : 0;
     }
 
     public function getRequestInfos()
     {
-        return $this->data['requestInfos'];
+        return $this->data['requestInfos'] ?? null;
     }
 
     public function getName()
@@ -47,7 +59,6 @@ class CommercetoolsDataCollector extends DataCollector
 
     public function reset()
     {
-        $this->profile = null;
+        $this->data = [];
     }
-
 }
