@@ -12,8 +12,8 @@ class Configuration implements ConfigurationInterface
 {
     public function getConfigTreeBuilder()
     {
-        $treeBuilder = new TreeBuilder();
-        $rootNode = $treeBuilder->root('commercetools');
+        $treeBuilder = new TreeBuilder('commercetools');
+        $rootNode = $this->getRootNode($treeBuilder, 'commercetools');
 
         $rootNode
             ->children()
@@ -139,5 +139,14 @@ class Configuration implements ConfigurationInterface
         ;
 
         return $treeBuilder;
+    }
+
+    private function getRootNode(TreeBuilder $treeBuilder, $name)
+    {
+        // BC layer for symfony/config 4.1 and older
+        if ( ! \method_exists($treeBuilder, 'getRootNode')) {
+            return $treeBuilder->root($name);
+        }
+        return $treeBuilder->getRootNode();
     }
 }
