@@ -7,20 +7,20 @@ namespace Commercetools\Symfony\ExampleBundle\Controller;
 
 
 use Commercetools\Core\Client;
-use Commercetools\Core\Error\InvalidArgumentException;
 use Commercetools\Core\Model\Customer\CustomerReference;
 use Commercetools\Core\Model\Product\ProductReference;
 use Commercetools\Core\Model\Review\Review;
 use Commercetools\Symfony\CtpBundle\Model\QueryParams;
 use Commercetools\Symfony\ExampleBundle\Model\Form\Type\AddReviewType;
 use Commercetools\Symfony\ReviewBundle\Manager\ReviewManager;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Component\Workflow\Exception\InvalidArgumentException;
 use Symfony\Component\Workflow\Registry;
 
-class ReviewController extends Controller
+class ReviewController extends AbstractController
 {
     /**
      * @var Client
@@ -59,6 +59,7 @@ class ReviewController extends Controller
     {
         $params = new QueryParams();
         $params->add('expand', 'customer');
+        $params->add('sort', 'createdAt desc');
         $reviews = $this->manager->getByProductId($request->getLocale(), $productId, $params);
 
         $reviewForm = $this->createForm(AddReviewType::class);
@@ -108,7 +109,7 @@ class ReviewController extends Controller
      * @param UserInterface|null $user
      * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
      */
-    public function UpdateReviewAction(Request $request, $reviewId, UserInterface $user = null)
+    public function updateReviewAction(Request $request, $reviewId, UserInterface $user = null)
     {
         if(is_null($user)){
             $this->addFlash('error', 'Do not allow anonymous reviews for now');
