@@ -10,20 +10,18 @@ use Commercetools\Core\Request\Customers\Command\CustomerChangeAddressAction;
 use Commercetools\Core\Request\Customers\Command\CustomerChangeEmailAction;
 use Commercetools\Core\Request\Customers\Command\CustomerSetFirstNameAction;
 use Commercetools\Core\Request\Customers\Command\CustomerSetLastNameAction;
-use Commercetools\Core\Request\Customers\CustomerByIdGetRequest;
-use Commercetools\Symfony\CustomerBundle\Entity\UserAddress;
-use Commercetools\Symfony\CustomerBundle\Entity\UserDetails;
+use Commercetools\Symfony\ExampleBundle\Entity\UserAddress;
+use Commercetools\Symfony\ExampleBundle\Entity\UserDetails;
 use Commercetools\Symfony\ExampleBundle\Model\Form\Type\AddressType;
 use Commercetools\Symfony\ExampleBundle\Model\Form\Type\UserType;
 use Commercetools\Symfony\CustomerBundle\Manager\CustomerManager;
-use Commercetools\Symfony\CustomerBundle\Security\User\User;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Security\Core\User\UserInterface;
 
 
-class UserController extends Controller
+class UserController extends AbstractController
 {
     /**
      * @var Client
@@ -44,19 +42,19 @@ class UserController extends Controller
         $this->manager = $manager;
     }
 
-    public function indexAction()
-    {
-        /**
-         * @var User $user
-         */
-        $user = $this->getUser();
-
-        return $this->render('ExampleBundle:catalog:index.html.twig',
-            [
-                'user' => $user
-            ]
-        );
-    }
+//    public function indexAction()
+//    {
+//        /**
+//         * @var User $user
+//         */
+//        $user = $this->getUser();
+//
+//        return $this->render('ExampleBundle:catalog:index.html.twig',
+//            [
+//                'user' => $user
+//            ]
+//        );
+//    }
 
     public function loginAction(Request $request)
     {
@@ -66,12 +64,10 @@ class UserController extends Controller
 
         $lastUsername = $authenticationUtils->getLastUsername();
 
-        return $this->render('ExampleBundle:user:login.html.twig',
-            [
-                'last_username' => $lastUsername,
-                'error' => $error
-            ]
-        );
+        return $this->render('ExampleBundle:user:login.html.twig', [
+            'last_username' => $lastUsername,
+            'error' => $error
+        ]);
     }
 
     public function detailsAction(Request $request, UserInterface $user)
@@ -98,10 +94,9 @@ class UserController extends Controller
                 ->setLastName(CustomerSetLastNameAction::of()->setLastName($lastName))
                 ->changeEmail(CustomerChangeEmailAction::ofEmail($email));
 
-
             try {
                 $customer = $customerBuilder->flush();
-            } catch (\Error $e){
+            } catch (\Exception $e){
                 $this->addFlash('error', $e->getMessage());
             }
 
@@ -110,22 +105,18 @@ class UserController extends Controller
             }
         }
 
-        return $this->render('ExampleBundle:User:user.html.twig',
-            [
-                'formDetails' => $form->createView()
-            ]
-        );
+        return $this->render('ExampleBundle:User:user.html.twig', [
+            'formDetails' => $form->createView()
+        ]);
     }
 
     public function addressBookAction(Request $request, UserInterface $user)
     {
         $customer = $this->manager->getById($request->getLocale(), $user->getId());
 
-        return $this->render('ExampleBundle:User:addressBook.html.twig',
-            [
-                'customer' => $customer
-            ]
-        );
+        return $this->render('ExampleBundle:User:addressBook.html.twig', [
+            'customer' => $customer
+        ]);
     }
 
     public function editAddressAction(Request $request, UserInterface $user, $addressId)
@@ -149,30 +140,29 @@ class UserController extends Controller
             $customerBuilder->flush();
         }
 
-        return $this->render(
-            'ExampleBundle:User:editAddress.html.twig',
-            [
-                'form_address' => $form->createView()
-            ]
-        );
+        return $this->render('ExampleBundle:User:editAddress.html.twig', [
+            'form_address' => $form->createView()
+        ]);
     }
 
-    protected function getCustomer(User $user)
-    {
-        if (!$user instanceof User){
-            throw new \InvalidArgumentException;
-        }
+//    protected function getCustomer(User $user)
+//    {
+//        if (!$user instanceof User){
+//            throw new \InvalidArgumentException;
+//        }
+//
+//        /**
+//         * @var Client $client
+//         */
+//        $client = $this->get('commercetools.client');
+//
+//        $request = CustomerByIdGetRequest::ofId($user->getId());
+//        $response = $request->executeWithClient($client);
+//
+//        $customer = $request->mapResponse($response);
+//
+//        return $customer;
+//    }
 
-        /**
-         * @var Client $client
-         */
-        $client = $this->get('commercetools.client');
-
-        $request = CustomerByIdGetRequest::ofId($user->getId());
-        $response = $request->executeWithClient($client);
-
-        $customer = $request->mapResponse($response);
-
-        return $customer;
-    }
+    public function signUpAction(){}
 }
