@@ -4,19 +4,30 @@ namespace Commercetools\Symfony\StateBundle\Command;
 
 use Commercetools\Symfony\StateBundle\Model\ProcessStates;
 use Commercetools\Symfony\StateBundle\Model\Repository\StateRepository;
-use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\DependencyInjection\Container;
 use Symfony\Component\Yaml\Yaml;
 
-class CommercetoolsWorkflowCommand extends ContainerAwareCommand
+class CommercetoolsWorkflowCommand extends Command
 {
+    /**
+     * @var StateRepository
+     */
     private $repository;
 
-    public function __construct(StateRepository $repository)
+    /**
+     * @var Container
+     */
+    private $container;
+
+
+    public function __construct(StateRepository $repository, Container $container)
     {
         parent::__construct();
         $this->repository = $repository;
+        $this->container = $container;
     }
 
     protected function configure()
@@ -34,9 +45,13 @@ class CommercetoolsWorkflowCommand extends ContainerAwareCommand
         $stateTypes = $helper->parse($states);
 
         $yaml = Yaml::dump($stateTypes, 100, 4);
-        $filename = $this->getContainer()->get('kernel')->getRootDir() . '/../config/packages/workflow.yaml';
+        $kernel = $this->container->get('kernel');
 
-        file_put_contents($filename, $yaml);
+        $filename = $kernel->getProjectDir() . '/config/packages/' . $kernel->getEnvironment() . '/workflow.yaml';
+
+        if ($kernel->getEnvironment() !== 'test') {
+            file_put_contents($filename, $yaml);
+        }
 
         $output->writeln('Configuration file saved successfully at ' . $filename);
     }
