@@ -8,8 +8,7 @@ namespace Commercetools\Symfony\StateBundle\DependencyInjection\Compiler;
 
 use Commercetools\Symfony\StateBundle\Cache\StateKeyResolver;
 use Commercetools\Symfony\StateBundle\EventListener\TransitionSubscriber;
-use Commercetools\Symfony\StateBundle\Model\CtpMarkingStore;
-use Commercetools\Symfony\StateBundle\Model\StateType;
+use Commercetools\Symfony\StateBundle\Model\CtpMarkingStore\CtpMarkingStore;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
@@ -34,20 +33,11 @@ class CtpStateMachinePass implements CompilerPassInterface
             $container->setDefinition($fullClassName, $newMarkingStore);
 
             $container->register('transition_listener.' . $stateMachineName, TransitionSubscriber::class)
-                ->addArgument(new Reference('Commercetools\Symfony\StateBundle\Model\\' . $stateMachineName . 'TransitionHandler'))
+                ->addArgument(new Reference('Commercetools\Symfony\StateBundle\Model\TransitionHandler\\' . $stateMachineName . 'TransitionHandler'))
                 ->addTag('kernel.event_listener', [
                     'event' => 'workflow.' . $stateMachineName . '.transition',
                     'method' => 'transitionSubject'
                 ]);
         }
-    }
-
-    private function getStateMachineType($type)
-    {
-        $types = array_combine(
-            array_map('current', StateType::TYPES),
-            array_keys(StateType::TYPES)
-        );
-        return $types[$type];
     }
 }
