@@ -6,7 +6,7 @@
 namespace Commercetools\Symfony\StateBundle\EventListener;
 
 
-use Commercetools\Symfony\StateBundle\Model\SubjectHandler;
+use Commercetools\Symfony\StateBundle\Model\TransitionHandler\SubjectHandler;
 use Symfony\Component\Workflow\Event\Event;
 
 class TransitionSubscriber
